requirements.txt
torch
timm
transformers<4.56
gradio
pillow
# python_orb_slam3
pillow-jxl-plugin